+++ app/src/main/java/fi/kroon/vadret/data/radar/net/RadarNetDataSource.kt
@@ -4,11 +4,15 @@ import fi.kroon.vadret.data.radar.model.Radar
 import io.reactivex.Single
 import retrofit2.Response
 import retrofit2.http.GET
+import retrofit2.http.Headers
 import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface RadarNetDataSource {
 
+    @Headers(
+        value = ["Cache-Control: public, max-stale=2419200"]
+    )
     @GET("/api/version/latest/area/sweden/product/comp/{year}/{month}/{date}")
     fun getRadar(
         @Path("year") year: String,