+++ app/src/main/res/raw/changelog.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added okhttp cache
+
 ## [1.2.8] -- 2020-05-08
 
 ### Fixes